+++ griptape/core/drivers/memory/memory_driver.py
