+++ griptape/core/drivers/prompt/base_prompt_driver.py
