+++ griptape/core/drivers/prompt/openai_prompt_driver.py
@@ -3,7 +3,7 @@ from typing import Optional
 import openai
 from attr import define, field, Factory
 from griptape.artifacts import TextOutput
-from griptape.drivers import BasePromptDriver
+from griptape.core.drivers import BasePromptDriver
 from griptape.tokenizers import TiktokenTokenizer
