+++ tests/unit/schemas/test_pipeline_schema.py
@@ -1,4 +1,4 @@
-from griptape.drivers import OpenAiPromptDriver
+from griptape.core.drivers import OpenAiPromptDriver
 from griptape.tokenizers import TiktokenTokenizer
 from griptape.tasks import PromptTask, ToolkitTask, BaseTask
 from griptape.structures import Pipeline
